added delete method for payment objects
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@5081 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ payment/classes/class.ilPaymentObject.php
@@ -123,13 +123,18 @@ class ilPaymentObject
 	}
 	function delete()
 	{
-		$query = "DELETE FROM payment_objects ".
-			"WHERE pobject_id = '".$this->getPobjectId()."'";
-
-		$this->db->query($query);
-
-		return true;
+		if($this->getPobjectId())
+		{
+			$query = "DELETE FROM payment_objects ".
+				"WHERE pobject_id = '".$this->getPobjectId()."'";
+			
+			$this->db->query($query);
+			
+			return true;
+		}
+		return false;
 	}
+
 	function update()
 	{
 		$query = "UPDATE payment_objects ".
